Refactor chatServiceHandler
